Detect overflow in line numbers in diff output.
* src/diff-state.h (diff_state): change in_lineno and out_lineno from
int to long, for consistency with strtol(), which is what we use to
parse numbers in the diff output.
(diff_state::in_line, diff_state::out_line): update return type to be
consistent.
* src/except.h: Add new exception class CsscInternalRangeException.
* src/diff-state.cc (get_num): all decimal conversions are now
performed through this function.  Correctly detect range errors in
the input.  Throw CsscInternalRangeException when this happens.
(diff_state::next_state): change line number variables from int to
long.  Use get_num() for conversion instead of strtol().  Eliminate
C-style casts.

+++ src/diff-state.h
@@ -68,7 +68,7 @@ private:
 public:
   diff_state(FILE *f, bool echo)
     : _state(diffstate::START),
-      in_lineno(0), out_lineno(0),
+      in_lineno(0L), out_lineno(0L),
       lines_left(0), change_left(0),
       echo_diff_output(echo),
       in(f),
@@ -86,8 +86,8 @@ public:
       return linebuf.c_str() + 2;
     }
 
-  int in_line() { return in_lineno; }
-  int out_line() { return out_lineno; }
+  long in_line() { return in_lineno; }
+  long out_line() { return out_lineno; }
 };
 
 #endif /* CSSC__DIFF_STATE_H */
